requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.1
python-dotenv==1.0.0
lxml==4.9.3
gunicorn==21.2.0 